requirements.txt
streamlit>=1.22.0
pandas>=1.3.5
numpy>=1.20.0
plotly>=5.10.0
statsmodels>=0.13.2
scikit-learn>=1.0.2
scipy>=1.7.3
matplotlib>=3.5.1
umap-learn>=0.5.3